backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers>=4.3.1
djangorestframework==3.17.1
gunicorn>=22.0.0
Pillow>=10.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.5
tzdata==2026.1
